refactor: create `pic` module
